Fix Warning double -> Float

+++ src/GameEngine/Entity/Component.cpp
@@ -271,23 +271,23 @@ ge::Vector2f ge::Collider::VectorIntersec(Vector2f const & uStart, Vector2f cons
 	Vector2f u = uEnd - uStart;
 	Vector2f v = vEnd - vStart;
 
-	double DotProduct = (u.x * v.y) - (v.x * u.y);
+	float DotProduct = (u.x * v.y) - (v.x * u.y);
 	if (DotProduct == 0)
 		return (Vector2f(-1, -1));
 	bool positive = DotProduct > 0;
 
 	Vector2f s = uStart - vStart;
-	double snumer = (u.x * s.y) - (s.x * u.y);
+	float snumer = (u.x * s.y) - (s.x * u.y);
 	if ((snumer < 0) == positive)
 		return (Vector2f(-1, -1));
 
-	double tnumer = (v.x * s.y) - (s.x * v.y);
+	float tnumer = (v.x * s.y) - (s.x * v.y);
 	if ((tnumer < 0) == positive)
 		return (Vector2f(-1, -1));
 	if (((snumer > DotProduct) == positive) || ((tnumer > DotProduct) == positive))
 		return (Vector2f(-1, -1));
 
-	double t = tnumer / DotProduct;
+	float t = tnumer / DotProduct;
 	Vector2f intersection = Vector2f(uStart.x + (t *u.x), uStart.y + (t *u.y));
 	return (intersection);
 }
@@ -297,10 +297,10 @@ bool ge::Collider::AABBCircleIntersecQuick(Vector2f const &topLeftAABB, Vector2f
 	Vector2f AABBCenter = Vector2f(topLeftAABB.x + AABBSize.x, topLeftAABB.y + AABBSize.y);
 	AABBCenter.x = AABBCenter.x - (AABBSize.x / 2);
 	AABBCenter.y = AABBCenter.y - (AABBSize.y / 2);
-	double AABBRadius = Vector2f(topLeftAABB.x - AABBCenter.x, topLeftAABB.y - AABBCenter.y).length();
-	double dx = circleCenter.x - AABBCenter.x;
-	double dy = circleCenter.y - AABBCenter.y;
-	double dist = sqrtf(dx * dx + dy * dy);
+	float AABBRadius = Vector2f(topLeftAABB.x - AABBCenter.x, topLeftAABB.y - AABBCenter.y).length();
+	float dx = circleCenter.x - AABBCenter.x;
+	float dy = circleCenter.y - AABBCenter.y;
+	float dist = sqrtf(dx * dx + dy * dy);
 	return (dist < AABBRadius + radius);
 }
